+++ apps/decodex/src/orchestrator/reconciliation/stalled/markers.rs
@@ -1,0 +1,24 @@
+use crate::{orchestrator::reconciliation::Path, state};
+
+pub(super) fn write_reconciliation_operation_marker_best_effort(
+	worktree_path: &Path,
+	run_id: &str,
+	attempt_number: i64,
+	current_operation: &str,
+) {
+	if let Err(error) = state::write_run_operation_marker_preserving_activity(
+		worktree_path,
+		run_id,
+		attempt_number,
+		current_operation,
+	) {
+		tracing::warn!(
+			?error,
+			run_id,
+			attempt_number,
+			current_operation,
+			worktree_path = %worktree_path.display(),
+			"Run operation marker write failed; continuing stalled-run reconciliation."
+		);
+	}
+}
