+++ apps/decodex/src/state/store/inputs.rs
@@ -1,0 +1,44 @@
+/// Input fields for recording a project-scoped external connector backoff.
+pub(crate) struct ConnectorBackoffInput<'a> {
+	pub(crate) project_id: &'a str,
+	pub(crate) connector: &'a str,
+	pub(crate) sync_phase: &'a str,
+	pub(crate) quota_class: &'a str,
+	pub(crate) reset_unix_epoch: i64,
+	pub(crate) reset_source: &'a str,
+	pub(crate) warning: &'a str,
+}
+
+/// Input fields for recording the latest review-policy checkpoint.
+pub(crate) struct ReviewPolicyCheckpointInput<'a> {
+	pub(crate) project_id: &'a str,
+	pub(crate) issue_id: &'a str,
+	pub(crate) run_id: &'a str,
+	pub(crate) attempt_number: i64,
+	pub(crate) phase: &'a str,
+	pub(crate) review_level: &'a str,
+	pub(crate) status: &'a str,
+	pub(crate) head_sha: &'a str,
+	pub(crate) nonclean_rounds: i64,
+	pub(crate) details_json: &'a str,
+}
+
+/// Input fields for looking up a review checkpoint by its reusable evidence key.
+pub(crate) struct ReviewCheckpointArtifactLookup<'a> {
+	pub(crate) project_id: &'a str,
+	pub(crate) issue_id: &'a str,
+	pub(crate) phase: &'a str,
+	pub(crate) review_level: &'a str,
+	pub(crate) head_sha: &'a str,
+}
+
+/// Input fields for recording the latest loop-guardrail checkpoint.
+pub(crate) struct LoopGuardrailCheckpointInput<'a> {
+	pub(crate) project_id: &'a str,
+	pub(crate) issue_id: &'a str,
+	pub(crate) reason: &'a str,
+	pub(crate) fingerprint: &'a str,
+	pub(crate) run_id: &'a str,
+	pub(crate) attempt_number: i64,
+	pub(crate) details_json: &'a str,
+}
